requirements.txt
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.12.3
certifi==2025.4.26
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
outcome==1.3.0.post0
packaging==25.0
PySocks==1.7.1
python-dotenv==1.1.0
python-telegram-bot==20.8
requests==2.31.0
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
